+++ ChangeLog.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- NEW : remove clone function *04/01/2022* - 1.27.0
 - NEW : Ajout extrafield sur catégorie pour faire apparaitre le tag catégorie *04/01/2022* - 1.26.0  
   -> lors du préremplissage des notes de saisie de temps sur les tâches   
 - NEW : Préremplissage des notes de saisie de temps sur les tâches *04/01/2022* - 1.25.0